Add LICENSE and README
In preparation for publishing, a LICENSE and a README have been added.

The README is simple enough; it just explains the rationale of this
program, as well as provides some pointers for hacking on the code for
any potential readers.

MIT is chosen for more permissive conditions than GPL (which is usually
my license of choice), since the goal here is for learning and education
purposes.